[ISSUE #9621]♻️Extract Tokio client inline tests into a sibling module
